use long long for length

+++ api.h
@@ -28,8 +28,8 @@ int crypto_sign_keypair(unsigned char *pk, unsigned char *sk);
  *
  * Returns 0 (success), -1 otherwise
  **************************************************/
-int crypto_sign(unsigned char *sm, unsigned long *smlen,
-                const unsigned char *m, unsigned long mlen, unsigned char *sk);
+int crypto_sign(unsigned char *sm, unsigned long long *smlen,
+                const unsigned char *m, unsigned long long mlen, unsigned char *sk);
 
 /*************************************************
  * Name:        crypto_sign_open
@@ -45,8 +45,8 @@ int crypto_sign(unsigned char *sm, unsigned long *smlen,
  *
  * Returns 0 if signed message could be verified correctly and -1 otherwise
  **************************************************/
-int crypto_sign_open(unsigned char *m, unsigned long *mlen,
-                     const unsigned char *sm, unsigned long smlen, const unsigned char *pk);
+int crypto_sign_open(unsigned char *m, unsigned long long *mlen,
+                     const unsigned char *sm, unsigned long long smlen, const unsigned char *pk);
 
 /*************************************************
  * Name:        crypto_remain_signatures
